mediaai-studio/requirements.txt
# MediaAI Corp — Dépendances
# Installez avec : pip install -r requirements.txt

# Vidéo
moviepy>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0

# Audio
gTTS>=2.5.0
# pyttsx3>=2.90  # optionnel, peut avoir des conflits

# Traduction
deep-translator>=1.11.0
langdetect>=1.0.9

# YouTube API
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
httplib2>=0.22.0

# E-Commerce
requests>=2.31.0
# woocommerce>=3.0.0  # optionnel
# pyactiveresource>=2.1.2  # pour Shopify (optionnel)

# Utilitaires
python-dotenv>=1.0.0
